refactor: remove unnecessary bubble edges class, keep code in a single file, add license reference
